requirements.txt
colr
nltk 
emoji
numpy
distro
pillow
pyttsx3
pdf2image
ansicolors
python-vlc
pytesseract